refactor: remove processGesture from proxy to handler
